+++ packages/sources/ncfx-test/src/endpoint/index.ts
@@ -1,2 +1,0 @@
-export { cryptoEndpoint as crypto } from './crypto'
-export { forexEndpoint as forex } from './forex'