+++ packages/sources/ncfx/src/index.ts
@@ -1,10 +1,15 @@
-import { expose } from '@chainlink/ea-bootstrap'
-import { makeExecute, makeWSHandler, endpointSelector } from './adapter'
-import * as endpoints from './endpoint'
-import { makeConfig, NAME } from './config'
-import { envDefaultOverrides } from './config/envDefaultOverrides'
+import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
+import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
+import { customSettings } from './config'
+import { crypto, forex } from './endpoint'
+import includes from './config/includes.json'
 
-const adapterContext = { name: NAME, envDefaultOverrides }
+export const adapter = new PriceAdapter({
+  name: 'NCFX',
+  endpoints: [crypto, forex],
+  defaultEndpoint: crypto.name,
+  customSettings,
+  includes,
+})
 
-const { server } = expose(adapterContext, makeExecute(), makeWSHandler(), endpointSelector)
-export { NAME, makeExecute, makeConfig, server, endpoints }
+export const server = (): Promise<ServerInstance | undefined> => expose(adapter)