+++ packages/composites/reference-transform/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/reference-transform-adapter
 
+## 1.2.59
+
+### Patch Changes
+
+- Updated dependencies [838c9d927]
+  - @chainlink/ea-bootstrap@2.25.2
+  - @chainlink/ea@1.3.58
+  - @chainlink/ea-reference-data-reader@1.3.26
+
 ## 1.2.58
 
 ### Patch Changes