+++ packages/sources/finage/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## v1.1.0
 
+## 1.6.13
+
+### Patch Changes
+
+- Updated dependencies [838c9d927]
+  - @chainlink/ea-bootstrap@2.25.2
+  - @chainlink/ea-test-helpers@1.4.1
+
 ## 1.6.12
 
 ### Patch Changes