+++ packages/sources/paxos/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @chainlink/paxos-adapter
 
+## 1.4.19
+
+### Patch Changes
+
+- Updated dependencies [838c9d927]
+  - @chainlink/ea-bootstrap@2.25.2
+  - @chainlink/ea-test-helpers@1.4.1
+
 ## 1.4.18
 
 ### Patch Changes